+++ app/src/main/java/org/ipforsmartobjects/apps/popularmovies/detail/MovieDetailActivity.java
@@ -41,8 +41,8 @@ public class MovieDetailActivity extends AppCompatActivity {
             // Create the detail fragment and add it to the activity
             // using a fragment transaction.
             Bundle arguments = new Bundle();
-            arguments.putInt(MovieDetailFragment.ARG_ITEM_ID,
-                    getIntent().getIntExtra(MovieDetailFragment.ARG_ITEM_ID, -1));
+            arguments.putLong(MovieDetailFragment.ARG_ITEM_ID,
+                    getIntent().getLongExtra(MovieDetailFragment.ARG_ITEM_ID, -1));
             MovieDetailFragment fragment = new MovieDetailFragment();
             fragment.setArguments(arguments);
             getSupportFragmentManager().beginTransaction()
